Refactor frontend build and simulation integration
- Split the monolithic inline Babel/React code into modular components (SatelliteMap, icons, helpers, index.js).
- Added package.json, webpack.config.js, and .babelrc to configure Webpack and Babel.
- Externalized custom styling into mnet/static/css/style.css and updated main.html to load the bundled JS and external CSS.
- Updated run_mn.py to launch webpack in watch mode automatically during simulation startup, with proper signal handling for cleanup.
