+++ emulation/mnet/static/css/style.css
@@ -1,0 +1,57 @@
+/* mnet/static/css/style.css */
+
+.status-box {
+    display: inline-block;
+    padding: 4px 8px;
+    margin: 2px;
+    font-family: monospace;
+    border-radius: 4px;
+    font-size: 0.9em;
+  }
+  
+  .chart-container {
+    margin: 20px 0;
+  }
+  
+  .satellite-map-container {
+    background-color: #f8f9fa;
+    border-radius: 8px;
+    padding: 15px;
+    margin-bottom: 20px;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  }
+  
+  .section-container {
+    background-color: white;
+    border-radius: 8px;
+    padding: 20px;
+    margin-bottom: 20px;
+    border: 1px solid #dee2e6;
+  }
+  
+  .router-link {
+    color: #0d6efd;
+    text-decoration: none;
+  }
+  
+  .router-link:hover {
+    text-decoration: underline;
+  }
+  
+  .section-header {
+    cursor: pointer;
+    user-select: none;
+  }
+  
+  .section-header:hover {
+    background-color: #f8f9fa;
+  }
+  
+  .collapse-icon {
+    transition: transform 0.3s ease;
+  }
+  
+  .collapsed .collapse-icon {
+    transform: rotate(-90deg);
+  }
+
