+++ emulation/mnet/static/js/.babelrc
@@ -1,0 +1,4 @@
+{
+    "presets": ["@babel/preset-env", "@babel/preset-react"]
+  }
+
