+++ emulation/mnet/static/js/package.json
@@ -1,0 +1,24 @@
+{
+    "name": "satnet-simulator",
+    "version": "1.0.0",
+    "description": "Satellite Simulator UI",
+    "main": "index.js",
+    "scripts": {
+      "build": "webpack --mode production",
+      "watch": "webpack --watch --mode development"
+    },
+    "dependencies": {
+      "react": "^17.0.2",
+      "react-dom": "^17.0.2",
+      "plotly.js": "^2.16.1"
+    },
+    "devDependencies": {
+      "@babel/core": "^7.21.0",
+      "@babel/preset-env": "^7.21.0",
+      "@babel/preset-react": "^7.18.6",
+      "babel-loader": "^8.2.5",
+      "webpack": "^5.75.0",
+      "webpack-cli": "^5.0.1"
+    }
+  }
+
