+++ emulation/mnet/static/js/src/components/icons/SatelliteIcon.js
@@ -1,0 +1,22 @@
+import React from 'react';
+
+const SatelliteIcon = ({ size = 12, color = 'currentColor' }) => (
+  <svg
+    width={size}
+    height={size}
+    viewBox="0 0 24 24"
+    fill={color}
+    stroke={color}
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+  >
+    <path d="M13 7L9 3L5 7l4 4" />
+    <path d="m17 11 4 4-4 4-4-4" />
+    <path d="m8 12 4 4 6-6-4-4Z" />
+    <path d="m16 8 3-3" />
+    <path d="M9 21a6 6 0 0 0-6-6" />
+  </svg>
+);
+
+export default SatelliteIcon;
