+++ emulation/mnet/static/js/src/index.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import SatelliteMap from './components/SatelliteMap';
+
+ReactDOM.render(<SatelliteMap />, document.getElementById('satellite-map'));
