+++ emulation/mnet/templates/main.html
@@ -4,64 +4,12 @@
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
   <title>SatNetMiniSim</title>
-
   <!-- Bootstrap CSS -->
-  <link
-    href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css"
-    rel="stylesheet"
-  />
-
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet" />
+  <!-- External CSS for custom styling -->
+  <link rel="stylesheet" href="/static/css/style.css" />
   <!-- Plotly for graphs -->
   <script src="https://cdn.plot.ly/plotly-latest.min.js"></script>
-
-  <!-- Add base styling -->
-  <style>
-    .status-box {
-      display: inline-block;
-      padding: 4px 8px;
-      margin: 2px;
-      font-family: monospace;
-      border-radius: 4px;
-      font-size: 0.9em;
-    }
-    .chart-container {
-      margin: 20px 0;
-    }
-    .satellite-map-container {
-      background-color: #f8f9fa;
-      border-radius: 8px;
-      padding: 15px;
-      margin-bottom: 20px;
-      box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-    }
-    .section-container {
-      background-color: white;
-      border-radius: 8px;
-      padding: 20px;
-      margin-bottom: 20px;
-      border: 1px solid #dee2e6;
-    }
-    .router-link {
-      color: #0d6efd;
-      text-decoration: none;
-    }
-    .router-link:hover {
-      text-decoration: underline;
-    }
-    .section-header {
-      cursor: pointer;
-      user-select: none;
-    }
-    .section-header:hover {
-      background-color: #f8f9fa;
-    }
-    .collapse-icon {
-      transition: transform 0.3s ease;
-    }
-    .collapsed .collapse-icon {
-      transform: rotate(-90deg);
-    }
-  </style>
 </head>
 <body class="bg-light">
   <div class="container">
@@ -124,11 +72,7 @@
 
           <!-- Ground Stations Section -->
           <div class="section-container">
-            <h3
-              class="section-header d-flex align-items-center"
-              data-bs-toggle="collapse"
-              data-bs-target="#ground-stations-content"
-            >
+            <h3 class="section-header d-flex align-items-center" data-bs-toggle="collapse" data-bs-target="#ground-stations-content">
               <span class="collapse-icon me-2">▼</span>
               Ground Stations
             </h3>
@@ -137,8 +81,7 @@
                 {% for entry in info["stations"] %}
                   <li class="mb-4">
                     <div class="d-flex align-items-center mb-2">
-                      <a href="{{ url_for('view_station', name=entry.name) }}"
-                         class="router-link me-3">
+                      <a href="{{ url_for('view_station', name=entry.name) }}" class="router-link me-3">
                         {{ entry.name }}
                       </a>
                       <span class="text-muted">{{ entry.defaultIP() }}</span>
@@ -168,11 +111,7 @@
 
           <!-- Routers Section -->
           <div class="section-container">
-            <h3
-              class="section-header d-flex align-items-center"
-              data-bs-toggle="collapse"
-              data-bs-target="#routers-content"
-            >
+            <h3 class="section-header d-flex align-items-center" data-bs-toggle="collapse" data-bs-target="#routers-content">
               <span class="collapse-icon me-2">▼</span>
               Routers ({{ info["routers"]|length }})
             </h3>
@@ -181,8 +120,7 @@
                 {% for entry in info["routers"] %}
                   <li class="mb-3">
                     <div class="d-flex align-items-center mb-2">
-                      <a href="{{ url_for('view_router', node=entry[0]) }}"
-                         class="router-link me-3">
+                      <a href="{{ url_for('view_router', node=entry[0]) }}" class="router-link me-3">
                         {{ entry[0] }}
                       </a>
                       <span class="text-muted">{{ entry[1] }}</span>
@@ -211,24 +149,15 @@
 
   <!-- Bootstrap JS Bundle with Popper -->
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"></script>
-
-  <!-- React & ReactDOM (development or production - your choice) -->
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/react/17.0.2/umd/react.development.js"></script>
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/react-dom/17.0.2/umd/react-dom.development.js"></script>
-
-  <!-- Babel for in-browser JSX transpiling -->
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/babel-standalone/6.26.0/babel.min.js"></script>
-
-  <!-- Single-file React code with SatelliteMap + icons + render call -->
-  <!-- Make sure /static/js/SatelliteMap.js contains everything (icons & ReactDOM.render) -->
-  <script type="text/babel" src="/static/js/SatelliteMap.js"></script>
-
+  <!-- Production React and ReactDOM -->
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/react/17.0.2/umd/react.production.min.js"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/react-dom/17.0.2/umd/react-dom.production.min.js"></script>
+  <!-- Load the bundled JS (built by Webpack) -->
+  <script src="/static/js/dist/bundle.js"></script>
   <!-- Connection Status Charts -->
   <script>
-    // Create the connection status charts using Plotly
     function createConnectionCharts() {
       {% if info["monitor_stable_nodes"] %}
-      // Stable nodes chart
       const stableTraceFail = {
         x: {{ info["stats_dates"]|tojson }},
         y: {{ info["stats_stable_fail"]|tojson }},
@@ -253,7 +182,6 @@
       Plotly.newPlot('stable-nodes-chart', [stableTraceFail, stableTraceOk], stableLayout);
       {% endif %}
 
-      // Dynamic nodes chart
       const dynamicTraceFail = {
         x: {{ info["stats_dates"]|tojson }},
         y: {{ info["stats_dynamic_fail"]|tojson }},
